Added __attribute__ format constraints on printf-like functions.

+++ liquidwar6/src/lib/cfg/cfg-xml.c
@@ -333,7 +333,7 @@ lw6cfg_read_key_value_xml_file (const char *filename,
 			      _x_
 			      ("parse error reading XML file \"%s\" at line %d: \"%s\""),
 			      filename,
-			      XML_GetCurrentLineNumber (parser),
+			      (int) XML_GetCurrentLineNumber (parser),
 			      XML_ErrorString (XML_GetErrorCode (parser)));
 #endif
 		}
