+++ liquidwar6/src/lib/ldr/ldr-bw.c
@@ -122,8 +122,8 @@ _lw6ldr_bw_read (_lw6ldr_image_bw_t * image, const char *png_file,
 		      lw6sys_log (LW6SYS_LOG_WARNING,
 				  _x_
 				  ("can't load B&W PNG file \"%s\", it is too big (size=%dx%d max=%dx%d)"),
-				  png_file, width, height, max_width,
-				  max_height);
+				  png_file, (int) width, (int) height,
+				  max_width, max_height);
 		      format_ok = 0;
 		    }
 
@@ -133,7 +133,8 @@ _lw6ldr_bw_read (_lw6ldr_image_bw_t * image, const char *png_file,
 				  _x_
 				  ("can't load B&W PNG file \"%s\", memory footprint is inconsistent, color_type=%d, rowbytes=%d, width=%d, step=%d, bit_depth=%d"),
 				  png_file, color_type,
-				  rowbytes, width, step, bit_depth);
+				  (int) rowbytes, (int) width, step,
+				  bit_depth);
 		      format_ok = 0;
 		    }
