+++ liquidwar6/src/lib/sys/sys-file.c
@@ -105,7 +105,7 @@ _read_bin (int *filesize, const char *filename)
 		  lw6sys_log (LW6SYS_LOG_INFO,
 			      _x_
 			      ("couldn't read all of file \"%s\", size was supposed to be %d but only read %d"),
-			      filename, size, read);
+			      filename, size, (int) read);
 #endif
 		}
 	      else if (read > size)
@@ -113,7 +113,7 @@ _read_bin (int *filesize, const char *filename)
 		  lw6sys_log (LW6SYS_LOG_WARNING,
 			      _x_
 			      ("inconsistency in file \"%s\", size is %d but read %d"),
-			      filename, size, read);
+			      filename, size, (int) read);
 		}
 	    }
 	}
@@ -222,7 +222,7 @@ lw6sys_write_file_content (const char *filename, const char *content)
 	{
 	  lw6sys_log (LW6SYS_LOG_WARNING,
 		      _x_ ("unable to write %d bytes in file \"%s\""),
-		      strlen (content), filename);
+		      (int) strlen (content), filename);
 	}
       fclose (f);
     }
